+++ src/Simplify.FluentNHibernate/Settings/ShowSqlOutputType.cs
@@ -1,0 +1,17 @@
+namespace Simplify.FluentNHibernate.Settings
+{
+	/// <summary>
+	/// Provides SQL command output type
+	/// </summary>
+	public enum ShowSqlOutputType
+	{
+		/// <summary>
+		/// The Console output
+		/// </summary>
+		Console,
+		/// <summary>
+		/// The Trace out[ut]
+		/// </summary>
+		Trace,
+	}
+}